Fix timeout decorator for async unary unary
